rename: 클래스 내부 파일 경로 수정
