+++ src/pages/Classes/components/Hero.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 
-import { curriculumImgUrl } from '@/pages/Classes/constants/imgUrl';
+import { curriculumImgUrl } from '../constants/imgUrl';
 
 function Hero() {
   const [showCurriculum, setShowCurriculum] = useState(false);
