+++ src/pages/Classes/components/form/ExamForm.tsx
@@ -3,8 +3,9 @@ import Quill from 'quill';
 import { FormEvent, useRef } from 'react';
 
 import QuillEditor from '@/components/QuillEditor';
-import { ExamInfo } from '@/types/classTypes';
 import base64ToFile from '@/utils/base64ToFile';
+
+import { ExamInfo } from '../../types/classTypes';
 
 function ExamForm({ examInfo }: { examInfo: ExamInfo }) {
   const { title, questions } = examInfo;
