+++ src/pages/Classes/index.tsx
@@ -1,7 +1,8 @@
 import Head from '@/components/Head';
-import CourseList from '@/pages/Classes/components/CourseList';
-import Hero from '@/pages/Classes/components/Hero';
-import MediaContents from '@/pages/Classes/components/MediaContents';
+
+import CourseList from './components/CourseList';
+import Hero from './components/Hero';
+import MediaContents from './components/MediaContents';
 
 function Classes() {
   return (
