+++ src/pages/Classes/types/classTypes.ts
